refactor(homepage): move DecapCms components off from admin/index
