+++ homepage/src/pages/admin/config.json
@@ -36,7 +36,11 @@
         },
         { "label": "Name", "name": "name", "widget": "string", "i18n": true },
         {
-          "label": "Page order", "name": "page_order", "widget": "number", "value_type": "int", "i18n": "duplicate"
+          "label": "Page order",
+          "name": "page_order",
+          "widget": "number",
+          "value_type": "int",
+          "i18n": "duplicate"
         },
         {
           "label": "Layout & Content",
